[sidecar] Single object for delivery
