+++ loadgen/client_test.go
@@ -237,7 +237,6 @@ func TestLoadGenForSidecar(t *testing.T) {
 			}
 			service, err := sidecar.New(sidecarConf)
 			require.NoError(t, err)
-			t.Cleanup(service.Close)
 			test.RunServiceAndServeForTest(t.Context(), t, service, sidecarServerConf)
 			// Start client
 			lgEnv.clientConf.Adapter.SidecarClient = &adapters.SidecarClientConfig{
@@ -284,7 +283,6 @@ func TestLoadGenForOrderer(t *testing.T) {
 			// Start sidecar.
 			service, err := sidecar.New(sidecarConf)
 			require.NoError(t, err)
-			t.Cleanup(service.Close)
 			test.RunServiceAndServeForTest(t.Context(), t, service, serverConfig)
 
 			// Start client
